% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{create_run_table}
\alias{create_run_table}
\title{Create *run* table}
\usage{
create_run_table(schema_name, table_name = "run", drop = FALSE,
  silent = FALSE, conn = getOption("pg.conn"), .log = getOption("pg.log",
  TRUE))
}
\arguments{
\item{schema_name}{character scalar}

\item{table_name}{character scalar}

\item{drop}{logical not yet supported.}

\item{silent}{logical if TRUE will wrap into silent `try`.}

\item{conn}{active connection to postgres database, by default `getOption("pg.conn")`.}

\item{.log}{logical default `getOption("pg.log",TRUE)` decides if call is logged using *logR*.}
}
\description{
*run* table generates run_id for `new_run_id` function from a sequence, it also stores R timestamp and db timestamp.
}

