% Generated by roxygen2: do not edit by hand
\name{new_run_id}
\alias{new_run_id}
\title{Generate new run id}
\usage{
new_run_id(conn = getOption("pg.conn"), .log = getOption("pg.log", TRUE))
}
\arguments{
\item{conn}{active connection to postgres database, by default `getOption("pg.conn")`.}

\item{.log}{logical default `getOption("pg.log",TRUE)` decides if call is logged using *logR*.}
}
\description{
Query data base for a new run id from sequence, storing also timestamp of the run id obtained.
}

