% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg-package.R
\docType{package}
\name{pg-package}
\alias{pg}
\alias{pg-package}
\title{pg: Polya-Gamma Distribution Sampler}
\description{
Provides access to a high performant random distribution sampler for the Polya-Gamma Distribution using either C++ headers for 'Rcpp' or 'RcppArmadillo' and 'R'.
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} [copyright holder]

}
\keyword{internal}
