% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg-catalog.R
\name{pg_catalog.pg_tables}
\alias{pg_catalog.pg_tables}
\title{Tables from PostgreSQL Catalogue}
\usage{
pg_catalog.pg_tables(src, ...)
}
\arguments{
\item{src}{Database connection.}

\item{...}{
  Arguments passed on to \code{\link[=pg_catalog]{pg_catalog}}
  \describe{
    \item{\code{pg_table}}{Character vector. Name of PostgreSQL catalogue table.}
  }}
}
\value{
Lazy table. Table of tables. Columns include schema and table name,
table owner.
}
\description{
Tables from PostgreSQL Catalogue
}
\examples{
\dontrun{
pg::pg_catalog.pg_tables(src) \%>\%
  dplyr::filter(!(schemaname \%in\% c("pg_catalog", "information_schema")))
}
}
