% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg-catalog.R
\name{pg_catalog.pg_views}
\alias{pg_catalog.pg_views}
\title{Views from PostgreSQL Catalogue}
\usage{
pg_catalog.pg_views(src, ...)
}
\arguments{
\item{src}{Database connection.}

\item{...}{
  Arguments passed on to \code{\link[=pg_catalog]{pg_catalog}}
  \describe{
    \item{\code{pg_table}}{Character vector. Name of PostgreSQL catalogue table.}
  }}
}
\value{
Lazy table.
}
\description{
Views from PostgreSQL Catalogue
}
