% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg-catalog.R
\name{pg_catalog}
\alias{pg_catalog}
\title{PostgreSQL Catalogue Table}
\usage{
pg_catalog(src, pg_table, ...)
}
\arguments{
\item{src}{PostgreSQL database connection.}

\item{pg_table}{Character vector. Name of PostgreSQL catalogue table.}

\item{...}{
  Arguments passed on to \code{\link[dplyr:tbl]{dplyr::tbl}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
Lazy catalogue table.
}
\description{
PostgreSQL Catalogue Table
}
