% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg-in-schema.R
\name{pg_in_schema}
\alias{pg_in_schema}
\title{PostgreSQL Schemas from Schema Tables}
\usage{
pg_in_schema(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[dplyr:compute]{dplyr::collect}}
  \describe{
    \item{\code{x}}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for more
details.}
  }}
}
\value{
Collected data frame with \code{in_schema} column of type
list-encapsulated \code{dbplyr_s} structure representing the database
schema-table.
}
\description{
Mutates a data frame by adding \code{in_schema} equal to the database schema-table
tuple based on the PostgreSQL schema name and table name.
}
\examples{
\dontrun{
pg::pg_catalog.pg_tables(src) \%>\%
  dplyr::filter(!(schemaname \%in\% c("pg_catalog", "information_schema"))) \%>\%
  pg::pg_in_schema()
}
}
