% Generated by roxygen2: do not edit by hand
\docType{package}
\name{pg-package}
\alias{pg-package}
\title{pg - postgres utilities for R}
\description{
Set of handy function when extensively working with postgres database. Includes transactional logging using *logR* package. Read `?pg` manual, use similarly to DBI functions.
}

