% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg-package.R
\docType{package}
\name{pg-package}
\alias{pg}
\alias{pg-package}
\title{pg: Polya Gamma Distribution Sampler}
\description{
Provides access to a series of highly performant random distribution samplers for the Polya Gamma Distribution as described by Polson, Scott, and Windle (2013) \href{https://arxiv.org/abs/1205.0310}{arXiv:1205.0310} using either 'C++' headers for 'Rcpp' or 'RcppArmadillo' and 'R'. The 'C++' header approach was developed to enable computations in Balamuta (2021) \url{https://www.ideals.illinois.edu/items/121209}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tmsalab.github.io/pg/}
  \item \url{https://github.com/tmsalab/pg}
  \item Report bugs at \url{https://github.com/tmsalab/pg/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} [copyright holder]

}
\keyword{internal}
