% Generated by roxygen2: do not edit by hand
\name{ppc}
\alias{ppc}
\title{prettyprint.char}
\usage{
ppc(trunc.char)
}
\arguments{
\item{trunc.char}{integer scalar passed as option value}
}
\description{
Wrapper to setup `options("datatable.prettyprint.char")` option. Affects data.table column width while printing.
}

